+++ src/main/java/com/kainos/ea/execption/DatabaseConnectionException.java
@@ -1,4 +1,4 @@
-package com.kainos.ea.client;
+package com.kainos.ea.execption;
 
 public class DatabaseConnectionException extends Throwable {
     public DatabaseConnectionException(Exception e) {
